requirements-production.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-dotenv>=1,<2
psycopg[binary]>=3.2,<4
psutil>=5.9,<8
